remove errors from creating new empty components
