+++ client/src/components/BudgetMonth.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import DetailList from './DetailList';
-import EnvelopeGroupList from './EnvelopeGroupList';
+import DetailList from './detailListComponents/DetailList';
+import EnvelopeGroupList from './envelopeGroupListComponents/EnvelopeGroupList';
 
 export default function BudgetMonth() {
   // This needs to get passed down to both Envelope and DetailList
