+++ client/src/components/Nav.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export default function Nav() {
+  return(
+    <div>
+    </div>
+  );
+}
