+++ client/src/components/detailListComponents/FundDetail.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function FundDetail() {
+  return(
+    <div>
+      {/* Detailed properties and calculations of the fund */}
+    </div>
+  );
+}
